+++ packages/ZbayLite/src/renderer/components/windows/Index.js
@@ -6,11 +6,11 @@ import Typography from '@material-ui/core/Typography'
 import Card from '@material-ui/core/Card'
 import CardContent from '@material-ui/core/CardContent'
 import CardMedia from '@material-ui/core/CardMedia'
+import LinearProgress from '@material-ui/core/LinearProgress'
 import { withStyles } from '@material-ui/core/styles'
 
 import ZcashIcon from '../../static/images/zcash/zcash-icon-fullcolor.svg'
 import WindowWrapper from '../ui/WindowWrapper'
-import NodeStatus from '../../containers/widgets/node/NodeStatus'
 
 const styles = theme => ({
   icon: {
@@ -41,7 +41,7 @@ export const Index = ({ classes, bootstrapping, bootstrappingMessage }) => (
           <Typography variant='body1' align='justify' gutterBottom>
             { bootstrapping ? bootstrappingMessage : 'Waiting for Zcash node.'}
           </Typography>
-          <NodeStatus />
+          <LinearProgress />
         </CardContent>
       </Grid>
     </Card>