+++ packages/ZbayLite/src/renderer/containers/pages/Channel.js
@@ -5,17 +5,26 @@ import { connect } from 'react-redux'
 import ChannelComponent from '../../components/pages/Channel'
 
 import channelHandlers from '../../store/handlers/channel'
+import channelsSelectors from '../../store/selectors/channels'
+
+export const mapStateToProps = state => ({
+  generalChannelId: channelsSelectors.generalChannelId(state)
+})
 
 export const mapDispatchToProps = dispatch => bindActionCreators({
   loadChannel: channelHandlers.epics.loadChannel
 }, dispatch)
 
 // TODO: after enzyme starts supporting hooks write tests
-const Channel = ({ loadChannel, loadChannels, match }) => {
+const Channel = ({ loadChannel, generalChannelId, match }) => {
   useEffect(() => {
-    loadChannel(match.params.id)
+    if (match.params.id === 'general') {
+      loadChannel(generalChannelId)
+    } else {
+      loadChannel(match.params.id)
+    }
   }, [match.params.id])
   return (<ChannelComponent />)
 }
 
-export default connect(null, mapDispatchToProps)(Channel)
+export default connect(mapStateToProps, mapDispatchToProps)(Channel)