+++ packages/ZbayLite/src/renderer/containers/windows/Vault.js
@@ -4,6 +4,7 @@ import { Redirect } from 'react-router'
 import { connect } from 'react-redux'
 
 import vaultSelectors from '../../store/selectors/vault'
+import vaultUnlockerSelectors from '../../store/selectors/vaultUnlocker'
 import nodeSelectors from '../../store/selectors/node'
 import vaultHandlers from '../../store/handlers/vault'
 import CreateVault from './CreateVault'
@@ -12,6 +13,7 @@ import UnlockVault from './UnlockVault'
 export const mapStateToProps = state => ({
   exists: vaultSelectors.exists(state),
   locked: vaultSelectors.locked(state),
+  unlocking: vaultUnlockerSelectors.unlocking(state),
   nodeConnected: nodeSelectors.isConnected(state)
 })
 
@@ -19,7 +21,7 @@ export const mapDispatchToProps = dispatch => bindActionCreators({
   loadVaultStatus: vaultHandlers.epics.loadVaultStatus
 }, dispatch)
 
-export const Vault = ({ loadVaultStatus, exists, locked, nodeConnected }) => {
+export const Vault = ({ loadVaultStatus, exists, unlocking, locked, nodeConnected }) => {
   useEffect(() => {
     loadVaultStatus()
   })
@@ -28,10 +30,10 @@ export const Vault = ({ loadVaultStatus, exists, locked, nodeConnected }) => {
   }
   if (!exists) {
     return <CreateVault />
-  } else if (locked) {
+  } else if (locked || unlocking) {
     return <UnlockVault />
   }
-  return <Redirect to='/main' />
+  return <Redirect to='/main/channel/general' />
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(Vault)